+++ Extension/src/common/messages/constants.ts
@@ -41,7 +41,7 @@ import {
     type Mv3LimitsCheckResult,
     type StaticLimitsCheckResult,
 } from '../../background/services/rules-limits/interface';
-import { type PageInitAppData } from '../../background/services/ui/main';
+import { type BlockingPageInitAppData, type PageInitAppData } from '../../background/services/ui/main';
 import { type ExportMessageResponse, type GetOptionsDataResponse } from '../../background/services/settings/types';
 import { type CreateEventListenerResponse } from '../../background/services/event';
 import { type FilterMetadata } from '../../background/api/filters/main';
@@ -91,10 +91,12 @@ export enum MessageType {
     OpenComparePage = 'openComparePage',
     ResetUserRulesForPage = 'resetUserRulesForPage',
     RemoveAllowlistDomain = 'removeAllowlistDomain',
-    AddAllowlistDomain = 'addAllowlistDomain',
+    AddAllowlistDomainForTabId = 'addAllowlistDomainForTabId',
+    AddAllowlistDomainForUrl = 'addAllowlistDomainForUrl',
     OnOpenFilteringLogPage = 'onOpenFilteringLogPage',
     GetFilteringLogData = 'getFilteringLogData',
     InitializeFrameScript = 'initializeFrameScript',
+    InitializeBlockingPageScript = 'initializeBlockingPageScript',
     OnCloseFilteringLogPage = 'onCloseFilteringLogPage',
     GetFilteringInfoByTabId = 'getFilteringInfoByTabId',
     SynchronizeOpenTabs = 'synchronizeOpenTabs',
@@ -372,10 +374,17 @@ export type SetEditorStorageContentMessage = {
     };
 };
 
-export type AddAllowlistDomainMessage = {
-    type: MessageType.AddAllowlistDomain;
+export type AddAllowlistDomainForTabIdMessage = {
+    type: MessageType.AddAllowlistDomainForTabId;
     data: {
         tabId: number;
+    };
+};
+
+export type AddAllowlistDomainForUrlMessage = {
+    type: MessageType.AddAllowlistDomainForUrl;
+    data: {
+        url: string;
     };
 };
 
@@ -493,6 +502,10 @@ export type DisableFiltersGroupMessage = {
 
 export type InitializeFrameScriptMessage = {
     type: MessageType.InitializeFrameScript;
+};
+
+export type InitializeBlockingPageScript = {
+    type: MessageType.InitializeBlockingPageScript;
 };
 
 export type SetConsentedFiltersMessage = {
@@ -764,8 +777,12 @@ export type MessageMap = {
         message: RemoveAllowlistDomainMessage;
         response: void;
     };
-    [MessageType.AddAllowlistDomain]: {
-        message: AddAllowlistDomainMessage;
+    [MessageType.AddAllowlistDomainForTabId]: {
+        message: AddAllowlistDomainForTabIdMessage;
+        response: void;
+    };
+    [MessageType.AddAllowlistDomainForUrl]: {
+        message: AddAllowlistDomainForUrlMessage;
         response: void;
     };
     [MessageType.LoadCustomFilterInfo]: {
@@ -907,6 +924,10 @@ export type MessageMap = {
     [MessageType.InitializeFrameScript]: {
         message: InitializeFrameScriptMessage;
         response: PageInitAppData;
+    };
+    [MessageType.InitializeBlockingPageScript]: {
+        message: InitializeBlockingPageScript;
+        response: BlockingPageInitAppData;
     };
 };