+++ Extension/src/pages/services/messenger.ts
@@ -43,7 +43,8 @@ import {
     type OpenSiteReportTabMessage,
     type ResetUserRulesForPageMessage,
     type RemoveAllowlistDomainMessage,
-    type AddAllowlistDomainMessage,
+    type AddAllowlistDomainForTabIdMessage,
+    type AddAllowlistDomainForUrlMessage,
     type GetFilteringInfoByTabIdMessage,
     type ClearEventsByTabIdMessage,
     type RefreshPageMessage,
@@ -104,7 +105,8 @@ class Messenger {
         this.resetUserRulesForPage = this.resetUserRulesForPage.bind(this);
         this.updateFilters = this.updateFilters.bind(this);
         this.removeAllowlistDomain = this.removeAllowlistDomain.bind(this);
-        this.addAllowlistDomain = this.addAllowlistDomain.bind(this);
+        this.addAllowlistDomainForTabId = this.addAllowlistDomainForTabId.bind(this);
+        this.addAllowlistDomainForUrl = this.addAllowlistDomainForUrl.bind(this);
     }
 
     /**
@@ -729,16 +731,32 @@ class Messenger {
     }
 
     /**
-     * Sends a message to the background page to add an allowlist domain.
+     * Sends a message to the background page to add an allowlist domain for a specific tab.
      *
      * @param tabId The ID of the tab.
      *
      * @returns Promise that resolves after the domain is added.
      */
-    async addAllowlistDomain(
-        tabId: AddAllowlistDomainMessage['data']['tabId'],
-    ): Promise<ExtractMessageResponse<MessageType.AddAllowlistDomain>> {
-        return this.sendMessage(MessageType.AddAllowlistDomain, { tabId });
+    async addAllowlistDomainForTabId(
+        tabId: AddAllowlistDomainForTabIdMessage['data']['tabId'],
+    ): Promise<ExtractMessageResponse<MessageType.AddAllowlistDomainForTabId>> {
+        return this.sendMessage(MessageType.AddAllowlistDomainForTabId, { tabId });
+    }
+
+    /**
+     * Sends a message to the background page to add an allowlist domain for a specific URL.
+     *
+     * Please note that after adding an allowlist domain, the tab will not be reloaded,
+     * it should be done separately later if needed.
+     *
+     * @param url The URL of the page.
+     *
+     * @returns Promise that resolves after the domain is added.
+     */
+    async addAllowlistDomainForUrl(
+        url: AddAllowlistDomainForUrlMessage['data']['url'],
+    ): Promise<ExtractMessageResponse<MessageType.AddAllowlistDomainForUrl>> {
+        return this.sendMessage(MessageType.AddAllowlistDomainForUrl, { url });
     }
 
     /**
@@ -1008,6 +1026,15 @@ class Messenger {
      */
     async initializeFrameScript(): Promise<ExtractMessageResponse<MessageType.InitializeFrameScript>> {
         return this.sendMessage(MessageType.InitializeFrameScript);
+    }
+
+    /**
+     * Sends a message to the background page to initialize the blocking page script.
+     *
+     * @returns Promise that resolves with the initialization data for the blocking page script.
+     */
+    async initializeBlockingPageScript(): Promise<ExtractMessageResponse<MessageType.InitializeBlockingPageScript>> {
+        return this.sendMessage(MessageType.InitializeBlockingPageScript);
     }
 
     /**