+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Updated Safebrowsing and blocked by rules blocking pages.
+
 ### Removed
 
 ### Fixed